Deduplicate Xorshift64 PRNG into shared library module
The identical Xorshift64 implementation was duplicated verbatim in
both `examples/lorawan_file_transfer/prng.rs` and
`tests/lorawan_frame_correctness.rs` (~45 lines each, ~90 total).

Move it to `src/prng.rs` as a public module. This aligns with the
architecture doc's principle of "seeded rand with explicit Rng
threading" -- having the deterministic PRNG in the core library makes
it available to all protocol implementors for reproducible simulations.

Both consumers now import `theatron::prng::Xorshift64` instead of
maintaining independent copies.
