+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod channel;
 pub mod metrics;
+pub mod prng;
 pub mod scheduler;
 pub mod time;
 pub mod traits;
