+++ tests/lorawan_frame_correctness.rs
@@ -4,41 +4,10 @@ use lorawan::parser::{DataHeader, DataPayload, MHDRAble, MType, PhyPayload};
 use lorawan_device::nb_device::radio::{Event, PhyRxTx, Response, RfConfig, RxQuality, TxConfig};
 use lorawan_device::nb_device::{Device, Event as DevEvent};
 use lorawan_device::{AppSKey, DevAddr, JoinMode, NewSKey, Timings};
-use rand_core::{Error, RngCore, impls};
+use rand_core::RngCore;
 
+use theatron::prng::Xorshift64;
 use theatron::types::Transmission;
-
-struct Xorshift64(u64);
-
-impl Xorshift64 {
-    fn new(seed: u64) -> Self {
-        Self(if seed == 0 { 1 } else { seed })
-    }
-}
-
-impl RngCore for Xorshift64 {
-    fn next_u32(&mut self) -> u32 {
-        self.next_u64() as u32
-    }
-
-    fn next_u64(&mut self) -> u64 {
-        let mut x = self.0;
-        x ^= x << 13;
-        x ^= x >> 7;
-        x ^= x << 17;
-        self.0 = x;
-        x
-    }
-
-    fn fill_bytes(&mut self, dest: &mut [u8]) {
-        impls::fill_bytes_via_next(self, dest);
-    }
-
-    fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), Error> {
-        self.fill_bytes(dest);
-        Ok(())
-    }
-}
 
 #[derive(Debug)]
 struct MinimalRadio {
